Refactor code to use double instead of float for square root calculation

+++ Khoá Học Thuật Toán T8-2024/01_phương_pháp_xử_lí_ước_số/test/DanhSachCacUoc.cpp
@@ -16,7 +16,7 @@ int main() {
         }
     }
     
-    float h = sqrt(n);
+    double h = sqrt(n);
     if ((int)h == h)
       uoc.pop_back();
