+++ Khoá Học Thuật Toán T8-2024/01_phương_pháp_xử_lí_ước_số/test/SLCU2.cpp
@@ -14,7 +14,7 @@ int main() {
         }
     }
 
-    float h = sqrt(n);
+    double h = sqrt(n);
     if ((int)h == h)
         dem--;
